testutils.go renamed to testutils_test.go
testutils.go imports the testing package. It shouldn't be builded otherwise "flag" package shows all flags from the "testing" package
